WIP - refactor; now zoekt-rs
