+++ Cargo.toml
@@ -1,8 +1,10 @@
 [workspace]
 members = [
 	"crates/hyperzoekt",
-	"crates/zoekt",
+	"crates/zoekt-rs",
 	"crates/zoekt-cli",
+	"crates/zoekt-bench",
+	"crates/zoekt-tools",
 ]
 
 # Use the edition-2021 dependency resolver semantics for workspace members.
