+++ crates/zoekt-bench/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "zoekt-bench"
+version = "0.0.1"
+edition = "2021"
+description = "Bench and microbench binaries for zoekt-rs"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "1.0"
+tempfile = "3"
+zoekt-rs = { path = "../zoekt-rs" }
+
+clap = { version = "4", features = ["derive"] }
+serde_json = "1"
+indicatif = "0.17"
+
+[dev-dependencies]
+criterion = "0.4"
