+++ crates/zoekt-bench/src/bin/zr-bench.rs
