+++ crates/zoekt-bench/src/bin/zr-compare-bench.rs
