+++ crates/zoekt-bench/src/bin/zr-query-bench.rs
