+++ crates/zoekt-cli/Cargo.toml
@@ -8,7 +8,7 @@ license = "Apache-2.0"
 [dependencies]
 anyhow = "1.0"
 clap = { version = "4", features = ["derive"] }
-zoekt-rs = { path = "../zoekt" }
+zoekt-rs = { path = "../zoekt-rs" }
 walkdir = "2"
 regex = "1"
 serde_json = "1"
