+++ crates/zoekt-rs/src/index.rs
