+++ crates/zoekt-rs/src/lib.rs
