+++ crates/zoekt-rs/src/shard.rs
