+++ crates/zoekt-rs/src/trigram.rs
