+++ crates/zoekt-rs/tests/index_builder_ignore_and_size.rs
