+++ crates/zoekt-rs/tests/smoke.rs
