+++ crates/zoekt-rs/tests/symbol_select.rs
