+++ crates/zoekt-rs/tests/symbol_unicode.rs
