+++ crates/zoekt-tools/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "zoekt-tools"
+version = "0.0.1"
+edition = "2021"
+description = "Developer tooling for zoekt-rs"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "1.0"
+regex = "1"
+zoekt-rs = { path = "../zoekt-rs" }
